requirements.txt
alembic==1.0.9
APScheduler==3.6.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Login==0.4.1
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
Flask-SSLify==0.1.5
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.9
MarkupSafe==1.1.1
PyMySQL==0.9.3
pyTelegramBotAPI==3.6.6
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2019.1
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.3
tzlocal==1.5.1
urllib3==1.24.1
Werkzeug==0.15.2
WTForms==2.2.1
